Apply flex layout to details and provide product context
