+++ manifest.json
@@ -21,6 +21,8 @@
     "vtex.breadcrumb": "1.x",
     "vtex.store-graphql": "2.x",
     "vtex.store-components": "3.x",
-    "vtex.styleguide": "9.x"
+    "vtex.styleguide": "9.x",
+    "vtex.flex-layout": "0.x",
+    "vtex.product-context": "0.x"
   }
 }
