+++ react/ProductDetailsLegacy.js
@@ -1,0 +1,3 @@
+import ProducstDetails from './legacy/ProductDetails'
+
+export default ProducstDetails
