+++ store/interfaces.json
@@ -1,15 +1,14 @@
 {
   "product-details": {
-    "required": [
+    "required": [],
+    "allowed": [
       "product-images",
       "product-name",
       "product-price",
       "sku-selector",
       "buy-button",
       "product-description",
-      "product-specifications"
-    ],
-    "allowed": [
+      "product-specifications",
       "breadcrumb",
       "shipping-simulator",
       "availability-subscriber",
@@ -17,9 +16,16 @@
       "product-highlights",
       "product-quantity-selector"
     ],
-    "component": "ProductDetails"
+    "component": "ProductDetailsLegacy"
   },
   "product-quantity-selector": {
     "component": "ProductQuantitySelector"
+  },
+  "product-details.layout": {
+    "component": "ProductDetails",
+    "composition": "children",
+    "allowed": [
+      "flex-layout"
+    ]
   }
 }
